+++ components/Tableau/embed_api/embed_api.js
@@ -1,0 +1,6 @@
+// eslint-disable-next-line no-unused-vars
+const tab_embed = typeof window !== 'undefined' ? require("./public.tableau.com_javascripts_api_tableau.embedding.3.latest.min.js") : null;
+
+// import tab_embed from './../api/public.tableau.com_javascripts_api_tableau.embedding.3.latest.min.js';
+
+export default tab_embed;
